overlays: Improve the i2c-rtc,i2c_csi_dsi option
The i2c_csi_dsi parameter of the i2c-rtc overlay (added for the CM4IO
board) causes the RTC devices to be probed on the I2C0 bus appearing
on GPIOs 44 and 45. However, it didn't enable the other nodes necessary
for it to work - "dtparam=i2c_vc=on" was also required.

Fix that.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>